share generator helpers in tools/internal/gen
